reverted back the compare function to activities
