+++ imports/ui/pages/activities/Activities.jsx
@@ -8,7 +8,6 @@ import { useTranslation } from 'react-i18next';
 import { Box, Button, Center, SimpleGrid, Wrap, WrapItem } from '@chakra-ui/react';
 
 import { StateContext } from '../../LayoutContainer';
-import { compareDatesForSort } from '../../utils/shared';
 import Loader from '../../components/Loader';
 import PublicActivityThumb from '../../components/PublicActivityThumb';
 
@@ -18,6 +17,13 @@ const publicSettings = Meteor.settings.public;
 const yesterday = moment().add(-1, 'days');
 
 const getFirstFutureOccurence = (occurence) => moment(occurence.endDate).isAfter(yesterday);
+function compareDatesForSort(a, b) {
+  const firstOccurenceA = a.datesAndTimes.find(getFirstFutureOccurence);
+  const firstOccurenceB = b.datesAndTimes.find(getFirstFutureOccurence);
+  const dateA = new Date(`${firstOccurenceA.startDate}T${firstOccurenceA.startTime}:00Z`);
+  const dateB = new Date(`${firstOccurenceB.startDate}T${firstOccurenceB.startTime}:00Z`);
+  return dateA - dateB;
+}
 
 function Activities({ activitiesList, processesList, isLoading, history }) {
   const { currentUser, currentHost, canCreateContent } = useContext(StateContext);
